requirements.txt
# Core dependencies
# FastAPI and web framework
fastapi>=0.121.3  # Python 3.12 support
uvicorn[standard]>=0.27.0  # Python 3.12 support

# Audio processing
mlx-whisper>=0.3.0  # Python 3.12 support on macOS

# Model Mirror
modelscope==1.29.0

# Core libraries
pydantic>=2.5.0,<3.0.0  # Pydantic v2 for Python 3.12
pyyaml>=6.0.1
python-multipart>=0.0.9
structlog>=24.1.0  # Python 3.12 compatible
aiofiles>=23.2.1  # Python 3.12 compatible
python-dotenv>=1.0.1

# Development dependencies
pytest>=8.0.0  # Python 3.12 support
pytest-asyncio>=0.23.5  # Python 3.12 compatible
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.27.0  # Python 3.12 compatible

# Code quality tools
black>=24.3.0  # Python 3.12 support
isort>=5.13.2  # Python 3.12 compatible
mypy>=1.8.0  # Python 3.12 support
